+++ kernel/expression/ExpressionCompoundVariable.hpp
@@ -52,10 +52,10 @@ class RexxQueue;
 
 class RexxCompoundVariable : public RexxVariableBase {
  public:
-  void *operator new(size_t, int);
+  void *operator new(size_t, size_t);
   inline void *operator new(size_t size, void *ptr) {return ptr;};
   inline void  operator delete(void *) { ; }
-  inline void  operator delete(void *, int) { ; }
+  inline void  operator delete(void *, size_t) { ; }
   inline void  operator delete(void *, void *) { ; }
 
   RexxCompoundVariable(RexxString *, size_t, RexxQueue *, size_t);
