+++ kernel/runtime/MemorySegment.cpp
@@ -385,7 +385,7 @@ void NormalSegmentSet::addDeadObject(DeadObject *object)
     else {
         /* calculate the dead chain          */
         /* and add that to the appropriate chain */
-        unsigned int deadChain = LengthToDeadPool(length);
+        size_t deadChain = LengthToDeadPool(length);
         subpools[deadChain].addSingle(object);
         /* we can mark this subpool as having items again */
         lastUsedSubpool[deadChain] = deadChain;
@@ -448,7 +448,7 @@ void NormalSegmentSet::addDeadObject(char *object, size_t length)
     else {
         /* calculate the dead chain          */
         /* and add that to the appropriate chain */
-        unsigned int deadChain = LengthToDeadPool(length);
+        size_t deadChain = LengthToDeadPool(length);
         subpools[deadChain].addSingle(new (object) DeadObject(length));
         /* we can mark this subpool as having items again */
         lastUsedSubpool[deadChain] = deadChain;
